requirements.txt
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
openai==1.57.0
PyPDF2==3.0.1
pdf2image==1.17.0
pytesseract==0.3.13
Pillow==11.0.0
python-dateutil==2.9.0
python-dotenv==1.0.0